Fix work content mobile width
